Added docstring
